code/pyproject.toml
[tool.poetry]
name = "todo"
version = "0.1.0"
description = ""
authors = ["Ch Saqib <chsaqib11223@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.110.0"
sqlmodel = "^0.0.16"
python-dotenv = "^1.0.1"
uvicorn = "^0.29.0"
psycopg2 = "^2.9.9"
pytest = "^8.1.1"
httpx = "^0.27.0"
aiokafka = "^0.10.0"
protobuf = "^5.27.1"




[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
